# responses/api-spec.yaml
openapi: 3.0.3
info:
  title: RecruitAI Backend API
  version: 0.1.0
  description: >
    MVP API for end-to-end Recruiting AI workflow (Rounds R1–R3) with JWT auth.
    Tech: Java/Spring Boot backend, PostgreSQL, Elasticsearch, Elastic Agent Builder.
servers:
  - url: http://localhost:8080/api
    description: Local dev
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with username/password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
        code: { type: string }
        details: {}
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    LoginResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    RefreshResponse:
      type: object
      properties:
        accessToken: { type: string }
    UserRole:
      type: string
      enum: [CANDIDATE, RECRUITER, ADMIN]
...